// projects/dxp-p1-003/scripts/provision-schema/src/types/DataverseMetadata.ts
export interface LocalizedLabel {
  readonly Label: string;
  readonly LanguageCode: number;
}

export interface LocalizedLabelSet {
  readonly LocalizedLabels: readonly LocalizedLabel[];
}
// ...
export interface OptionMetadata {
  readonly Value: number;
  readonly Label: LocalizedLabelSet;
}
// ...
export interface GlobalOptionSetPayload {
  readonly '@odata.type': 'Microsoft.Dynamics.CRM.OptionSetMetadata';
  readonly Name: string;
  readonly DisplayName: LocalizedLabelSet;
  readonly Description: LocalizedLabelSet;
  readonly IsGlobal: true;
  readonly OptionSetType: 'Picklist';
  readonly Options: readonly OptionMetadata[];
}
// ...
export interface ODataCollectionResponse<T> {
  readonly value: readonly T[];
  readonly '@odata.nextLink'?: string;
}

export interface PublisherRecord {
  readonly publisherid: string;
  readonly friendlyname: string;
  readonly customizationprefix: string;
}

export interface SolutionRecord {
  readonly solutionid: string;
  readonly uniquename: string;
  readonly version: string;
}

export interface GlobalOptionSetRecord {
  readonly MetadataId: string;
  readonly Name: string;
}